+++ charbot/events.py
@@ -25,13 +25,16 @@
 """Event handling for Charbot."""
 import json
 import re
+import sys
+import traceback
 from datetime import datetime, timedelta, timezone
 
 import discord
 from discord import Color, Embed
+from discord.ext import commands
 from discord.ext.commands import Cog
 
-from main import CBot
+from bot import CBot
 
 
 def sensitive_embed(message: discord.Message, used: set[str]) -> discord.Embed:
@@ -201,6 +204,65 @@ class Events(Cog):
                 re.MULTILINE | re.IGNORECASE,
             ):
                 await message.delete()
+
+    @Cog.listener()
+    async def on_command_error(self, ctx, error):
+        """Trigger when an error is raised while invoking a command.
+
+        Parameters
+        ----------
+        self : Events
+            The Events cog.
+        ctx: commands.Context
+            The context used for command invocation.
+        error: commands.CommandError
+            The Exception raised.
+        """
+        # This prevents any commands with local handlers being
+        # handled here in on_command_error.
+        if hasattr(ctx.command, "on_error"):
+            return
+
+        # This prevents any cogs with an overwritten
+        # cog_command_error being handled here.
+        cog: Cog = ctx.cog
+        if cog and (
+            # skipcq: PYL-W0212
+            cog._get_overridden_method(cog.cog_command_error)
+            is not None
+        ):
+            return
+
+        ignored = (commands.CommandNotFound,)
+
+        # Allows us to check for original exceptions
+        # raised and sent to CommandInvokeError.
+        # If nothing is found. We keep the exception passed to on_command_error.
+        error = getattr(error, "original", error)
+
+        # Anything in ignored will return and prevent anything happening.
+        if isinstance(error, ignored):
+            return
+
+        if isinstance(error, commands.DisabledCommand):
+            await ctx.send(f"{ctx.command} has been disabled.")
+
+        elif isinstance(error, commands.NoPrivateMessage):
+            try:
+                await ctx.author.send(f"{ctx.command} can not be used in Private Messages.")
+            except discord.HTTPException:
+                pass
+
+        # For this error example we check to see where it came from...
+        elif isinstance(error, commands.BadArgument):
+            await ctx.send("Bad Argument.")
+
+        else:
+            # All other Errors not returned come here.
+            # And we can just print the default TraceBack.
+            assert error is not None  # skipcq: BAN-B101
+            print(f"Ignoring exception in command {ctx.command}:", file=sys.stderr)
+            traceback.print_exception(type(error), error, error.__traceback__, file=sys.stderr)
 
 
 async def setup(bot: CBot):
