+++ charbot/sudoku/block.py
