+++ charbot/sudoku/cell.py
