+++ charbot/sudoku/column.py
