+++ charbot/sudoku/row.py
